+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.1.1 (2023-06-28)
 
 - create and test Core classes
-- update and use quiltcore.yaml
+- update and use config.yaml
 - implement get, put, and verify
 
 ## 0.1.0 (2023-06-16)